refactor(dkg): consolidate test files one per source file
Consolidate scattered test files so each source .go file has exactly
one corresponding _test.go file. No test cases were lost during
consolidation - they were moved between files.

Changes:
- Merge vote_justification_test.go, vote_prepare_test.go into vote_test.go
- Merge queue_justification_test.go into queue_test.go
- Merge dkg_handler_threshold_test.go into dkg_handler_internal_test.go
- Merge begin_dealing_test.go, dkg_process_test.go into dkg_dealing_test.go,
  dkg_finalization_internal_test.go, and dkg_rewards_internal_test.go
- Merge dkg_svc_dealing_guards, process_deals, justifications, should_deal,
  reprocess into dkg_svc_dealing_test.go
- Merge dkg_svc_decrypt, decrypt_full, worker, resume, old_cc into
  dkg_svc_test.go
- Merge dkg_svc_registration_guards into dkg_svc_registration_test.go
- Split dkg_svc_full_path_test.go into respective svc test files
- Rename dkg_lifecycle_internal_test.go to dkg_lifecycle_test.go
- Remove TestVerifyVoteExtension_BeforeV200 (tests removed V200 bypass)
- Remove vote_prepare_test.go (tests removed V200 height check logic)
- Add TestPrepareVotes_NoVEHeight for new ConsensusParams-based behavior
