+++ client/x/dkg/keeper/dkg_svc_finalization_test.go
@@ -4,8 +4,11 @@ import (
 	"context"
 	"testing"
 
+	ethtypes "github.com/ethereum/go-ethereum/core/types"
 	"github.com/stretchr/testify/require"
+	"go.uber.org/mock/gomock"
 
+	dkgtestutil "github.com/piplabs/story/client/x/dkg/testutil"
 	"github.com/piplabs/story/client/x/dkg/types"
 )
 
@@ -139,4 +142,81 @@ func TestHandleDKGFinalization_DuplicateAcquire(t *testing.T) {
 	got, err := k.stateManager.GetSession(5)
 	require.NoError(t, err)
 	require.Equal(t, types.PhaseDealing, got.Phase, "lock dedup should prevent processing")
+}
+
+// --- Full path tests merged from dkg_svc_full_path_test.go ---
+
+// TestHandleDKGFinalization_FullPath tests the full finalization path with
+// mocked kernel and contract clients.
+
+func TestHandleDKGFinalization_FullPath(t *testing.T) {
+	ctx := context.Background()
+
+	ctrl := gomock.NewController(t)
+
+	sm, err := NewStateManager(t.TempDir())
+	require.NoError(t, err)
+
+	cc := []byte("finalize-cc")
+	mockKernel := dkgtestutil.NewMockKernelServiceClient(ctrl)
+	mockContract := dkgtestutil.NewMockDKGContractClient(ctrl)
+
+	router := NewKernelRouter(nil, nil)
+	router.RegisterClient(cc, mockKernel)
+
+	k := &Keeper{
+		stateManager:     sm,
+		kernelRouter:     router,
+		contractClient:   mockContract,
+		validatorEVMAddr: testValidatorAddr,
+		enclaveType:      [32]byte{0x02},
+	}
+
+	resetDKGSvcRound()
+	defer resetDKGSvcRound()
+
+	// Create session in PhaseDealing
+	session := &types.DKGSession{
+		Round:          3,
+		Phase:          types.PhaseDealing,
+		CodeCommitment: cc,
+	}
+	require.NoError(t, sm.CreateSession(ctx, session))
+
+	// Kernel FinalizeDKG returns finalization data
+	mockKernel.EXPECT().FinalizeDKG(gomock.Any(), gomock.Any()).Return(
+		&types.FinalizeDKGResponse{
+			ParticipantsRoot: make([]byte, 32),
+			GlobalPubKey:     []byte("global-pub"),
+			Signature:        []byte("sig"),
+			PublicCoeffs:     [][]byte{[]byte("c1"), []byte("c2")},
+			PubKeyShare:      []byte("share"),
+		}, nil,
+	)
+
+	// Contract Finalize returns success
+	mockContract.EXPECT().Finalize(
+		gomock.Any(),
+		uint32(3),
+		gomock.Any(), // enclaveType
+		gomock.Any(), // participantsRoot
+		gomock.Any(), // globalPubKey
+		gomock.Any(), // publicCoeffs
+		gomock.Any(), // pubKeyShare
+		gomock.Any(), // signature
+	).Return(&ethtypes.Receipt{Status: ethtypes.ReceiptStatusSuccessful}, nil)
+
+	dkgNetwork := &types.DKGNetwork{
+		Round:        3,
+		Stage:        types.DKGStageFinalization,
+		ActiveValSet: []string{testValidatorAddr},
+	}
+
+	k.handleDKGFinalization(ctx, dkgNetwork)
+
+	got, err := sm.GetSession(3)
+	require.NoError(t, err)
+	require.Equal(t, types.PhaseFinalized, got.Phase)
+	require.Equal(t, []byte("global-pub"), got.GlobalPubKey)
+	require.Equal(t, []byte("sig"), got.SigFinalizeNetwork)
 }
